+++ Ray-Tracing/src/RayTrace/Material.java
@@ -1,16 +1,16 @@
 package RayTrace;
 
 class Material{
-     float diff_r, diff_g, diff_b;
-     float spec_r, spec_g, spec_b;
-     float reflect_r, reflect_g, reflect_b;
-     float phong;
-     float trans;
+     double diff_r, diff_g, diff_b;
+     double spec_r, spec_g, spec_b;
+     double reflect_r, reflect_g, reflect_b;
+     double phong;
+     double trans;
     
-    public Material(float diff_r, float diff_g, float diff_b,
-                    float spec_r, float spec_g, float spec_b,
-                    float reflect_r, float reflect_g, float reflect_b,
-                    float phong, float trans){
+    public Material(double diff_r, double diff_g, double diff_b,
+                    double spec_r, double spec_g, double spec_b,
+                    double reflect_r, double reflect_g, double reflect_b,
+                    double phong, double trans){
         this.diff_r = diff_r;
         this.diff_g = diff_g;
         this.diff_b = diff_b;
